Allow for different sources for specs and resources.
For the .NET implementation (from an offline email):
Is it possible to distinguish specifications and resources (such as embedded.css and visibility-toggler.js)?
The C# classes of Concordion.NET cover .NET specifics. For example the EmbeddedResourceSource loads executable specifications with the help of FixtureAssembly.GetManifestResourceStream(). This is quite different from loading resources, which is done with ResourceManager.GetString().

Also closes 2 input streams that weren't being closed.