Tidied up SpecificationLocator changes.
The previous change ensured the SpecificationLocator didn't change, so that we didn't break the ExcelExtension.
However, we now have a breaking change to the Source interface, and the ExcelExtension is being updated for Concordion 2.0.
This change adds a new method to SpecificationLocator and deprecates the existing method.
While it is a breaking change, it's much cleaner than the previous solution and more workable moving forward.