Add outer example listener.
In order for the parallel run extension to wait for each example to finish requires a listener for the "outer" example (which includes all commands in a specification not inside an example command) in addition to the existing example listener (which is only triggered for example commands).

Needed for concordion/concordion-parallel-run-extension#3.