Only run Concordion fixtures from the run command.
The run command checks for fixtures matching the specification that are named the same as the specification, or with Test or Fixture appended to the specification name.
Prior to this change, it would run matching classes regardless of whether they were Concordion fixtures.

The code now checks that the class is a Concordion fixture before running it.
By default, JUnit3 and JUnit4 fixtures are checked. For additional integrations, subclass the TestFrameworkProvider interface and follow the instructions in the
TestFrameworkProvider javadoc to integrate it.